csbackend/csmysite/requirements.txt
asgiref==3.7.2
distlib==0.3.8
distro-info==1.0
Django==5.0.3
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-filter==24.1
djangorestframework==3.14.0
environs==11.0.0
filelock==3.13.1
Markdown==3.6
marshmallow==3.21.1
packaging==24.0
platformdirs==4.2.0
PyMySQL==1.1.0
pypi==2.1
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.4.4
virtualenv==20.25.1
